commands: `spack load --list` alias for `spack find --loaded`
See #25249 and https://github.com/spack/spack/pull/27159#issuecomment-958163679.
This adds `spack load --list` as an alias for `spack find --loaded`.  The new command is
not as powerful as `spack find --loaded`, as you can't combine it with all the queries or
formats that `spack find` provides.  However, it is more intuitively located in the command
structure in that it appears in the output of `spack load --help`.

The idea here is that people can use `spack load --list`  for simple stuff but fall back to
`spack find --loaded` if they need more.

- [x] add help to `spack load --list` that references `spack find`
- [x] factor some parts of `spack find` out to be called from `spack load`
- [x] add shell integration
- [x] add shell test
- [x] update completion
- [x] update docs